requirements.txt
pillow
nltk
numpy
mkdocs
mkdocs-material
mkdocs-include-markdown-plugin
mkdocstrings